Fix type conversion warnings in K&R code
The additional flags started giving warnings and errors of floats and
ints being converted into doubles.

+++ k_r/chap_1/ex3_f2c.c
@@ -7,7 +7,7 @@
 #include <stdlib.h>
 
 int main(){
-    float fahr, celsius;
+    double fahr, celsius;
     int lower,upper, step;
 
     lower = -40;   // Lower limit of temperature table
@@ -19,7 +19,7 @@ int main(){
     while (fahr <= upper) {
 	celsius = 5.0 * ( fahr-32.0) / 9.0;
 	printf("%3.0f\t=>\t%6.1f\n",fahr,celsius);
-	fahr = fahr + step;
+	fahr = fahr + (double) step;
     }
     exit(0);
 }
