+++ k_r/chap_1/ex4_c2f.c
@@ -7,19 +7,19 @@
 #include <stdlib.h>
 
 int main(){
-    float fahr, celsius;
+    double fahr, celsius;
     int lower,upper, step;
 
     lower = -40;   // Lower limit of temperature table
     upper = 240; // Upper limit
     step = 20;
 
-    celsius = lower;
+    celsius = (double) lower;
     printf("Celsius to Fahrenheit Conversion\n");
     while (celsius <= upper) {
 	fahr = (( 9.0 / 5.0 ) * celsius ) +32.0;
 	printf("%3.0f\t=>\t%6.1f\n",celsius,fahr);
-	celsius = celsius + step;
+	celsius = celsius + (double) step;
     }
     exit(0);
 }
